+++ activerecord/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Fix support for `query_cache: false` in `database.yml`.
+
+    `query_cache: false` would no longer entirely disable the Active Record query cache.
+
+    *zzak*
+
 *   Fix marshalling of unsaved associated records in 7.1 format.
 
     The 7.1 format would only marshal associated records if the association was loaded.